+++ Changelog.md
@@ -4,6 +4,8 @@
 - Added Recipe Sorting
 - Added Refill Mouse Item to Smart Pickup
 - Added Fix Scroll direction to Fixed UI
+- Move Stack Trashed into Better Trash
+- Added Trash the Trash to Better Trash Slot
 - Fixed Quick with Frame Skip Off
 
 ## v0.8.2